requirements.txt
certifi==2019.6.16
chardet==3.0.4
facebook-sdk==3.1.0
idna==2.8
nltk==3.4.3
numpy==1.16.4
pandas==0.24.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
matplotlib==3.1.0
wordcloud==1.5.0
seaborn==0.9.0
